Moved interface *IUiService* to a separate file to improve the code readability
