+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+## [0.3.2] - 2020-04-18
+
+- Moved interface *IUiService* to a separate file to improve the code readability
+
 ## [0.3.1] - 2020-02-15
 
 - Updated dependency packages
